requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
httpx>=0.27.0
# Optional: only needed for XEvents reader
# pyodbc>=5.1.0
